feat: add build options to the route building (#101)
When building the documents on the dev server or in a preview server setting the build can pass along the request from express to allow the templates to be customized based on things like query params.

An example use case would be the 'help' text to show the partial and option information. Instead of needing to hide it with css, the template can use the request params to determine if the help information should even be rendered.