fix: Fix issue with uncaught exceptions from timers. (#215)
The timer execution was adding a promise .then handler but not handling exceptions.

This caused the error to be added to processes' unhandled exceptions even when the promise.all was also catching the error.

This change wraps the timer stopping and the triggering of the plugin hook into an async function. The Promise.all can correctly correlate the hooks while still getting accurate time execution.